backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy>=2.0.36,<3.0
Flask-SQLAlchemy==3.1.1
cryptography==41.0.7
reportlab==4.0.4
